// packages/sdk-pivot-query-client/src/data-handling/types.ts
export type PivotTreeNode<ContentType = any> = TreeNode & {
  // formatted value to show or React component
  content?: null | string | ContentType;
  contentType?: string;
  // metadata type (rows | columns | measures)
  metadataType?: string;
  // predefined user type (subTotal | grandTotal)
  userType?: string;
  // table type (corner | rows | columns | data)
  tableType?: string;
  // jaql panel field index
  jaqlIndex?: number;
  // jaql measure panel field index for last level column items
  measureJaqlIndex?: number;
  // represents path from root to current node (used in sorting)
  measurePath?: { [key: string]: string };
  // reference to parent PivotTreeNode
  parent?: PivotTreeNode<ContentType>;
  // reference to related PivotTreeNode (for sub/grand totals for example)
  master?: PivotTreeNode<ContentType>;
  // sorting direction
  dir?: string | null;
  // databars feature for measure
  databars?: boolean;
  // state of cell
  state?: {
    isSelected?: boolean;
    isDrilled?: boolean;
  };
  // cell data store
  store?: { domReadyDefer?: Defer; [key: string]: any };
};